requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
anyio==3.7.1
asttokens==2.4.1
attrs==23.2.0
azure-core==1.30.1
azure-functions==1.18.0
azure-identity==1.15.0
azure-storage-blob==12.19.1
bcrypt==4.1.2
caerus==0.1.9
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cryptography==41.0.7
cycler==0.12.1
dnspython==2.4.2
email-validator==2.1.0.post1
executing==2.0.1
fastapi==0.105.0
fastapi-cache2==0.2.1
fastapi-sso==0.9.1
findpeaks==2.6.1
fonttools==4.47.0
frozenlist==1.4.1
geojson==2.5.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
icecream==2.1.3
idna==3.6
iniconfig==2.0.0
isodate==0.6.1
joblib==1.3.2
kiwisolver==1.4.5
matplotlib==3.8.2
msal==1.28.0
msal-extensions==1.1.0
multidict==6.0.4
numpy==1.26.3
oauthlib==3.2.2
packaging==23.2
pandas==2.1.4
paramiko==3.4.0
peakdetect==1.1
pendulum==3.0.0
pgvector==0.2.4
pillow==10.2.0
pluggy==1.3.0
portalocker==2.8.2
psycopg==3.1.17
psycopg-binary==3.1.17
psycopg-pool==3.2.1
pycparser==2.21
pydantic==1.10.13
Pygments==2.17.2
PyJWT==2.8.0
PyNaCl==1.5.0
pyowm==3.3.0
pyparsing==3.1.1
PySocks==1.7.1
pytest==7.4.4
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
scipy==1.11.4
seaborn==0.13.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
sshtunnel==0.4.0
starlette==0.27.0
time-machine==2.13.0
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
uvicorn==0.24.0.post1
wget==3.2
xarray==2023.12.0
yarl==1.9.4